Conflicts:
	src/java/rexos/mas/data/Position.java
	src/java/rexos/mas/service_agent/ServiceAgent.java

+++ src/java/rexos/mas/logistics_agent/behaviours/GetPartsInfo.java
@@ -100,7 +100,7 @@ public class GetPartsInfo extends ReceiveOnceBehaviour {
 				int type = 3;
 				for(Part part : parts) {
 					partParameters.put(new Part(part.getType(), id++),
-							new Position(x++, 1, 3, new Part(type++, id + x)));
+							new Position((double)x++, 1.0, 3.0, new Part(type++, id + x)));
 				}
 
 				KnowledgeDBClient client = KnowledgeDBClient.getClient();
